Created vigenereTable(); Chngd "tablet" to "table"
Created a private method called vigenereTable(). This is the code that
creates the 2d char array representing the polyalpha cipher. It was in
both the encrypt and decrypt methods and is now on it's own. Changed the
array name from "tablet" to "table" and increased the array's scope
outside of the methods and made it private.